+++ build.clj
@@ -5,7 +5,7 @@
   clojure -T:build deploy
 
   Run tests via:
-  clojure -X:test
+  clojure -X:build test
 
   For more information, run:
 
@@ -15,7 +15,7 @@
             [deps-deploy.deps-deploy :as dd]))
 
 (def lib 'org.scicloj/clay)
-(def version "2-beta23")
+(def version "2-beta25")
 
 (def snapshot (str version "-SNAPSHOT"))
 (def class-dir "target/classes")
@@ -45,30 +45,30 @@
     [:tag (str "v" version)]]])
 
 (defn- jar-opts [opts]
-(let [version (if (:snapshot opts) snapshot version)]
-  (assoc opts
-         :lib lib   :version version
-         :jar-file  (format "target/%s-%s.jar" lib version)
-         :basis     (b/create-basis {})
-         :class-dir class-dir
-         :target    "target"
-         :src-dirs  ["src"]
-         :pom-data  (pom-template version))))
+  (let [version (if (:snapshot opts) snapshot version)]
+    (assoc opts
+           :lib lib   :version version
+           :jar-file  (format "target/%s-%s.jar" lib version)
+           :basis     (b/create-basis {})
+           :class-dir class-dir
+           :target    "target"
+           :src-dirs  ["src"]
+           :pom-data  (pom-template version))))
 
 (defn ci "Run the CI pipeline of tests (and build the JAR)." [opts]
-(test opts)
-(b/delete {:path "target"})
-(let [opts (jar-opts opts)]
-  (println "\nWriting pom.xml...")
-  (b/write-pom opts)
-  (println "\nCopying source...")
-  (b/copy-dir {:src-dirs ["resources" "src"] :target-dir class-dir})
-  (println "\nBuilding" (:jar-file opts) "...")
-  (b/jar opts))
-opts)
+  (test opts)
+  (b/delete {:path "target"})
+  (let [opts (jar-opts opts)]
+    (println "\nWriting pom.xml...")
+    (b/write-pom opts)
+    (println "\nCopying source...")
+    (b/copy-dir {:src-dirs ["resources" "src"] :target-dir class-dir})
+    (println "\nBuilding" (:jar-file opts) "...")
+    (b/jar opts))
+  opts)
 
 (defn deploy "Deploy the JAR to Clojars." [opts]
-(let [{:keys [jar-file] :as opts} (jar-opts opts)]
-  (dd/deploy {:installer :remote :artifact (b/resolve-path jar-file)
-              :pom-file (b/pom-path (select-keys opts [:lib :class-dir]))}))
-opts)
+  (let [{:keys [jar-file] :as opts} (jar-opts opts)]
+    (dd/deploy {:installer :remote :artifact (b/resolve-path jar-file)
+                :pom-file (b/pom-path (select-keys opts [:lib :class-dir]))}))
+  opts)