+++ test/index_generated_test.clj
@@ -14,15 +14,7 @@
 
 
 (def
- var3_line40
- ["(require '[scicloj.clay.v2.api :as clay])\n(clay/make! {:source-path \"notebooks/index.clj\"})"])
-
-
-(def var4_line42 nil)
-
-
-(def
- var5_line64
+ var3_line110
  (->>
   [["June 10th 2023" "An early overview - babashka-conf" "HvhMsv3iVGM"]
    ["Dec. 1st 2023"
@@ -46,23 +38,23 @@
   (into [:table])))
 
 
-(def var6_line92 nil)
+(def var4_line138 nil)
 
 
 (def
- var7_line138
+ var5_line184
  ["(do (require '[scicloj.clay.v2.api :as clay])\n    (clay/make! {:single-form '~form-before-caret\n                 :source-path [\"~file-path\"]}))"])
 
 
-(def var8_line141 nil)
+(def var6_line187 nil)
 
 
 (def
- var9_line146
+ var7_line192
  ["(do (require '[scicloj.clay.v2.api :as clay])\n    (clay/make! {:source-path [\"~file-path\"]}))"])
 
 
-(def var10_line148 nil)
+(def var8_line194 nil)
 
 
 (ns
@@ -79,33 +71,33 @@
   [clojure.test :refer [deftest is]]))
 
 
-(def var12_line171 nil)
+(def var10_line217 nil)
 
 
 (def
- var13_line172
+ var11_line218
  (kind/hiccup
   [:div
    {:style {:background "#efe9e6", :border-style :solid}}
    [:ul [:li "one"] [:li "two"] [:li "three"]]]))
 
 
-(def var14_line180 nil)
+(def var12_line226 nil)
 
 
 (def
- var15_line181
+ var13_line227
  (->
   {:x (range 5), :y (repeatedly 5 rand)}
   tc/dataset
   (tc/set-dataset-name "my dataset")))
 
 
-(def var16_line186 nil)
+(def var14_line232 nil)
 
 
 (def
- var17_line187
+ var15_line233
  (->
   (toydata/iris-ds)
   (hanami/plot
@@ -120,139 +112,139 @@
     :=mark-opacity 0.2})))
 
 
-(def var18_line198 nil)
+(def var16_line244 nil)
 
 
-(def var19_line200 (require '[scicloj.clay.v2.api :as clay]))
+(def var17_line246 (require '[scicloj.clay.v2.api :as clay]))
 
 
-(def var20_line202 nil)
+(def var18_line248 nil)
 
 
 (def
- var21_line209
+ var19_line255
  (comment
   (clay/make! {:format [:html], :source-path "notebooks/index.clj"})))
 
 
-(def var22_line213 nil)
+(def var20_line259 nil)
 
 
 (def
- var23_line215
+ var21_line261
  (comment (clay/make! {:source-path "notebooks/index.clj"})))
 
 
-(def var24_line218 nil)
+(def var22_line264 nil)
 
 
 (def
- var25_line222
+ var23_line268
  (comment
   (clay/make! {:source-path "notebooks/index.clj", :show false})))
 
 
-(def var26_line226 nil)
+(def var24_line272 nil)
 
 
 (def
- var27_line229
+ var25_line275
  (comment
   (clay/make!
    {:source-path "notebooks/index.clj",
     :favicon "notebooks/favicon.ico"})))
 
 
-(def var28_line233 nil)
+(def var26_line279 nil)
 
 
 (def
- var29_line237
+ var27_line283
  (comment
   (clay/make!
    {:source-path ["notebooks/slides.clj" "notebooks/index.clj"],
     :show false})))
 
 
-(def var30_line242 nil)
+(def var28_line288 nil)
 
 
 (def
- var31_line247
+ var29_line293
  (comment
   (clay/make!
    {:source-path ["notebooks/slides.clj" "notebooks/index.clj"],
     :live-reload true})))
 
 
-(def var32_line253 nil)
+(def var30_line299 nil)
 
 
 (def
- var33_line257
+ var31_line303
  (comment
   (clay/make!
    {:source-path "notebooks/index.clj", :single-form '(+ 1 2)})))
 
 
-(def var34_line261 nil)
+(def var32_line307 nil)
 
 
-(def var35_line265 (comment (clay/make! {:single-form '(+ 1 2)})))
+(def var33_line311 (comment (clay/make! {:single-form '(+ 1 2)})))
 
 
-(def var36_line268 nil)
+(def var34_line314 nil)
 
 
-(def var37_line271 (comment (clay/make! {:single-value 3})))
+(def var35_line317 (comment (clay/make! {:single-value 3})))
 
 
-(def var38_line274 nil)
+(def var36_line320 nil)
 
 
 (def
- var39_line278
+ var37_line324
  (comment
   (clay/make!
    {:single-value 3333,
     :post-process (fn [html] (-> html (str/replace #"3333" "4444")))})))
 
 
-(def var40_line284 nil)
+(def var38_line330 nil)
 
 
 (def
- var41_line287
+ var39_line333
  (comment
   (clay/make!
    {:source-path "notebooks/index.clj", :hide-ui-header true})))
 
 
-(def var42_line291 nil)
+(def var40_line337 nil)
 
 
 (def
- var43_line294
+ var41_line340
  (comment
   (clay/make!
    {:source-path "notebooks/index.clj", :hide-info-line true})))
 
 
-(def var44_line298 nil)
+(def var42_line344 nil)
 
 
 (def
- var45_line303
+ var43_line349
  (comment
   (clay/make!
    {:format [:quarto :html], :source-path "notebooks/index.clj"})))
 
 
-(def var46_line307 nil)
+(def var44_line353 nil)
 
 
 (def
- var47_line314
+ var45_line360
  (comment
   (clay/make!
    {:format [:quarto :html],
@@ -260,31 +252,31 @@
     :run-quarto false})))
 
 
-(def var48_line319 nil)
+(def var46_line365 nil)
 
 
 (def
- var49_line325
+ var47_line371
  (comment
   (clay/make!
    {:format [:quarto :html], :source-path "notebooks/slides.clj"})))
 
 
-(def var50_line329 nil)
+(def var48_line375 nil)
 
 
 (def
- var51_line335
+ var49_line381
  (comment
   (clay/make!
    {:format [:quarto :revealjs], :source-path "notebooks/slides.clj"})))
 
 
-(def var52_line339 nil)
+(def var50_line385 nil)
 
 
 (def
- var53_line345
+ var51_line391
  (comment
   (clay/make!
    {:format [:quarto :html],
@@ -293,11 +285,11 @@
     {:highlight-style :nord, :format {:html {:theme :journal}}}})))
 
 
-(def var54_line351 nil)
+(def var52_line397 nil)
 
 
 (def
- var55_line361
+ var53_line407
  (comment
   (require
    '[scicloj.clay.v2.quarto.highlight-styles
@@ -312,21 +304,21 @@
      :format {:html {:theme quarto.themes/journal}}}})))
 
 
-(def var56_line369 nil)
+(def var54_line415 nil)
 
 
 (def
- var57_line374
+ var55_line420
  (comment
   (clay/make!
    {:base-source-path "notebooks/", :source-path "index.clj"})))
 
 
-(def var58_line378 nil)
+(def var56_line424 nil)
 
 
 (def
- var59_line380
+ var57_line426
  (comment
   (clay/make!
    {:format [:quarto :html],
@@ -341,11 +333,11 @@
     :clean-up-target-dir true})))
 
 
-(def var60_line393 nil)
+(def var58_line439 nil)
 
 
 (def
- var61_line395
+ var59_line441
  (comment
   (clay/make!
    {:format [:quarto :html],
@@ -356,11 +348,11 @@
     :clean-up-target-dir true})))
 
 
-(def var62_line407 nil)
+(def var60_line453 nil)
 
 
 (def
- var63_line409
+ var61_line455
  (comment
   (clay/make!
    {:format [:quarto :html],
@@ -373,25 +365,25 @@
     :clean-up-target-dir true})))
 
 
-(def var64_line422 nil)
+(def var62_line468 nil)
 
 
-(def var65_line425 (comment (clay/browse!)))
+(def var63_line471 (comment (clay/browse!)))
 
 
-(def var66_line428 nil)
+(def var64_line474 nil)
 
 
 (def
- var67_line442
+ var65_line492
  (comment (clay/make-hiccup {:source-path "notebooks/index.clj"})))
 
 
-(def var68_line445 nil)
+(def var66_line495 nil)
 
 
 (def
- var69_line508
+ var67_line558
  (->>
   ["purple" "darkgreen" "brown"]
   (mapcat
@@ -405,19 +397,19 @@
   kind/fragment))
 
 
-(def var70_line516 (->> (range 3) kind/fragment))
+(def var68_line566 (->> (range 3) kind/fragment))
 
 
-(def var71_line519 nil)
+(def var69_line569 nil)
 
 
 (def
- var72_line528
+ var70_line578
  (kind/fn {:x 1, :y 2} #:kindly{:f (fn [{:keys [x y]}] (+ x y))}))
 
 
 (def
- var73_line533
+ var71_line583
  (kind/fn
   {:my-video-src
    "https://file-examples.com/storage/fe58a1f07d66f447a9512f1/2017/04/file_example_MP4_480_1_5MG.mp4"}
@@ -427,59 +419,59 @@
             (kind/video {:src my-video-src}))}))
 
 
-(def var74_line538 nil)
+(def var72_line588 nil)
 
 
-(def var75_line540 (kind/fn [+ 1 2]))
+(def var73_line590 (kind/fn [+ 1 2]))
 
 
-(def var76_line543 nil)
+(def var74_line593 nil)
 
 
 (def
- var77_line545
+ var75_line595
  (kind/fn {:kindly/f (fn [{:keys [x y]}] (+ x y)), :x 1, :y 2}))
 
 
-(def var78_line551 nil)
+(def var76_line601 nil)
 
 
 (def
- var79_line554
+ var77_line604
  (kind/fn
   {:x (range 3), :y (repeatedly 3 rand)}
   #:kindly{:f tc/dataset}))
 
 
 (def
- var80_line559
+ var78_line609
  (kind/fn [tc/dataset {:x (range 3), :y (repeatedly 3 rand)}]))
 
 
 (def
- var81_line564
+ var79_line614
  (kind/fn {:kindly/f tc/dataset, :x (range 3), :y (repeatedly 3 rand)}))
 
 
-(def var82_line569 nil)
+(def var80_line619 nil)
 
 
-(def var83_line577 (delay (Thread/sleep 500) (+ 1 2)))
+(def var81_line627 (delay (Thread/sleep 500) (+ 1 2)))
 
 
-(def var84_line581 nil)
+(def var82_line631 nil)
 
 
 (def
- var85_line587
+ var83_line637
  (kind/hiccup [:img {:src "notebooks/images/Clay.svg.png"}]))
 
 
-(def var86_line590 (kind/image {:src "notebooks/images/Clay.svg.png"}))
+(def var84_line640 (kind/image {:src "notebooks/images/Clay.svg.png"}))
 
 
 (def
- var87_line593
+ var85_line643
  (kind/vega-lite
   {:data {:url "notebooks/datasets/iris.csv"},
    :mark "rule",
@@ -494,106 +486,106 @@
    :background "floralwhite"}))
 
 
-(def var88_line605 nil)
+(def var86_line655 nil)
 
 
-(def var89_line618 (+ 1 2))
+(def var87_line668 (+ 1 2))
 
 
-(deftest test90_line620 (is (> var89_line618 2.9)))
+(deftest test88_line670 (is (> var87_line668 2.9)))
 
 
-(deftest test91_line623 (is (> var89_line618 2.9)))
+(deftest test89_line673 (is (> var87_line668 2.9)))
 
 
-(deftest test92_line625 (is (> var89_line618 2.9)))
+(deftest test90_line675 (is (> var87_line668 2.9)))
 
 
-(def var93_line627 nil)
+(def var91_line677 nil)
 
 
 (def
- var94_line634
+ var92_line684
  (kindly/hide-code
   (kind/code
    "(kind/test-last [> 2.9])\n\n^kind/test-last\n[> 2.9]\n\n(kindly/check > 2.9)")))
 
 
-(def var95_line643 nil)
+(def var93_line693 nil)
 
 
 (def
- var96_line653
+ var94_line703
  (kind/table
   {:column-names ["A" "B" "C"], :row-vectors [[1 2 3] [4 5 6]]}
   {:class "table-responsive", :style {:background "#f8fff8"}}))
 
 
-(def var97_line658 nil)
+(def var95_line708 nil)
 
 
 (def
- var98_line689
+ var96_line739
  (kindly/hide-code
   (kindly/merge-options! {:code-and-value :horizontal})
   false))
 
 
-(def var99_line693 (+ 1 2))
+(def var97_line743 (+ 1 2))
 
 
-(def var100_line695 (+ 3 4))
+(def var98_line745 (+ 3 4))
 
 
-(def var101_line697 nil)
+(def var99_line747 nil)
 
 
 (def
- var102_line699
+ var100_line749
  (kindly/hide-code
   (kindly/merge-options! {:code-and-value :vertical})
   false))
 
 
-(def var103_line703 (+ 1 2))
+(def var101_line753 (+ 1 2))
 
 
-(def var104_line705 (+ 3 4))
+(def var102_line755 (+ 3 4))
 
 
-(def var105_line707 nil)
+(def var103_line757 nil)
 
 
 (def
- var106_line709
+ var104_line759
  (kindly/hide-code
   (kindly/merge-options! {:style {:background-color "#ccddee"}})
   false))
 
 
-(def var107_line713 (kind/hiccup [:div [:p "hello"]]))
+(def var105_line763 (kind/hiccup [:div [:p "hello"]]))
 
 
-(def var108_line717 nil)
+(def var106_line767 nil)
 
 
-(def var109_line720 (tc/dataset {:x (range 3)}))
+(def var107_line770 (tc/dataset {:x (range 3)}))
 
 
-(def var110_line722 nil)
+(def var108_line772 nil)
 
 
-(def var111_line724 (kind/hiccup [:div (tc/dataset {:x (range 3)})]))
+(def var109_line774 (kind/hiccup [:div (tc/dataset {:x (range 3)})]))
 
 
-(def var112_line728 nil)
+(def var110_line778 nil)
 
 
 (def
- var113_line730
+ var111_line780
  (kindly/hide-code
   (kindly/merge-options! {:style {:background-color nil}})
   false))
 
 
-(def var114_line734 (kind/hiccup [:div [:p "hello"]]))
+(def var112_line784 (kind/hiccup [:div [:p "hello"]]))